+++ examples/with-next/next.config.js
@@ -8,7 +8,14 @@ const withTM = require('next-transpile-modules')
  * @type {import('next').NextConfig}
  **/
 module.exports = withPlugins(
-  [withTM(['@twind/core', '@twind/preset-autoprefix', '@twind/preset-tailwind', '@twind/with-next'])],
+  [
+    withTM([
+      '@twind/core',
+      '@twind/preset-autoprefix',
+      '@twind/preset-tailwind',
+      '@twind/with-next',
+    ]),
+  ],
   {
     /* regular next.js config options here */
     reactStrictMode: true,